Add MeshServiceProtocol and Windows implementation
