+++ Package.swift
@@ -2,13 +2,17 @@
 
 import PackageDescription
 
+var supportedPlatforms: [SupportedPlatform] = [
+    .iOS(.v16),
+    .macOS(.v13)
+]
+#if os(Windows)
+supportedPlatforms.append(.windows(.v10))
+#endif
+
 let package = Package(
     name: "bitchat",
-    platforms: [
-        .iOS(.v16),
-        .macOS(.v13),
-        .windows(.v10)
-    ],
+    platforms: supportedPlatforms,
     products: [
         .executable(
             name: "bitchat",
