+++ bitchat/Services/WindowsBluetoothMeshService.swift
@@ -1,0 +1,68 @@
+#if os(Windows)
+import Foundation
+import WinSDK
+
+class WindowsBluetoothMeshService: NSObject, MeshServiceProtocol {
+    weak var delegate: BitchatDelegate?
+    var myPeerID: String = UUID().uuidString
+
+    // MARK: - Service Control
+    func startServices() {
+        // TODO: Implement scanning and advertising using WinRT APIs
+    }
+
+    // MARK: - Messaging
+    func sendBroadcastAnnounce() {
+        // Windows implementation pending
+    }
+
+    func sendMessage(_ content: String, mentions: [String], channel: String?) {
+        // Windows implementation pending
+    }
+
+    func sendPrivateMessage(_ content: String, to peerID: String, recipientNickname: String, messageID: String?) {
+    }
+
+    func sendEncryptedChannelMessage(_ content: String, mentions: [String], channel: String, channelKey: SymmetricKey, messageID: String?, timestamp: Date?) {
+    }
+
+    func sendReadReceipt(_ receipt: ReadReceipt, to peerID: String) {
+    }
+
+    func sendChannelMetadata(_ metadata: ChannelMetadata) {
+    }
+
+    func sendChannelLeaveNotification(_ channel: String) {
+    }
+
+    func announcePasswordProtectedChannel(_ channel: String, creatorID: String?, keyCommitment: String?) {
+    }
+
+    func sendChannelRetentionAnnouncement(_ channel: String, enabled: Bool) {
+    }
+
+    func sendChannelPasswordUpdate(_ password: String, channel: String, newCommitment: String, to peerID: String) {
+    }
+
+    func sendChannelKeyVerifyRequest(_ request: ChannelKeyVerifyRequest, to peers: [String]) {
+    }
+
+    func sendChannelKeyVerifyResponse(_ response: ChannelKeyVerifyResponse, to peerID: String) {
+    }
+
+    func emergencyDisconnectAll() {
+    }
+
+    func getNoiseService() -> NoiseEncryptionService {
+        return NoiseEncryptionService()
+    }
+
+    func getPeerFingerprint(_ peerID: String) -> String? { nil }
+
+    func getPeerNicknames() -> [String : String] { [:] }
+
+    func getPeerRSSI() -> [String : NSNumber] { [:] }
+
+    func getFingerprint(for peerID: String) -> String? { nil }
+}
+#endif
